removed count neighbours loop as it was useless. only one loop is needed now. I get 22 fps on 1000 boids

+++ src/utils.cpp
@@ -19,11 +19,11 @@ sf::Vector2f vec2::vecFromDegree(const float deg) {
 }
 
 float vec2::distanceBetweenPoints(const sf::Vector2f& v1, const sf::Vector2f& v2) {
-    return sqrt((v1.x - v2.x) * (v1.x - v2.x) + (v1.y - v2.y) * (v1.y - v2.y));
+    return sqrtf((v1.x - v2.x) * (v1.x - v2.x) + (v1.y - v2.y) * (v1.y - v2.y));
 }
 
 float vec2::getMagnitude(const sf::Vector2f& vec) {
-    return sqrt(vec.x * vec.x + vec.y * vec.y);
+    return sqrtf(vec.x * vec.x + vec.y * vec.y);
 }
 
 sf::Vector2f vec2::normalize(const sf::Vector2f& vec) {
